refactor: 131 - move function into different file
