+++ src/components/catalog/product-list.ts
@@ -3,6 +3,7 @@ import { createProductCard } from './card/card';
 import type { ProductProjectionPagedQueryResponse } from '@commercetools/platform-sdk';
 import { CATALOG } from '../../pages/catalog/constants';
 import { CustomEventEmitter } from '../../utils/event-emitter';
+import { handleSearchResults } from '../../helpers/handle-search-results';
 
 export const searchEventEmitter = new CustomEventEmitter();
 
@@ -14,23 +15,7 @@ export function createProductList(products: ProductProjectionPagedQueryResponse)
     wrapper.append(card);
   }
 
-  searchEventEmitter.subscribe('search', (data) => {
-    if (typeof data === 'object' && data !== null && 'results' in data && Array.isArray(data.results)) {
-      wrapper.replaceChildren();
-      if (data.results.length === 0) {
-        createDiv({
-          classes: CATALOG.noResults,
-          parent: wrapper,
-          text: 'No results found',
-        });
-      }
-
-      for (const product of data.results) {
-        const card = createProductCard(product);
-        wrapper.append(card);
-      }
-    }
-  });
+  searchEventEmitter.subscribe('search', (data) => handleSearchResults(wrapper, data));
 
   return wrapper;
 }
